merge_entities: extract turn_into_redirection
